requirements.txt
flask
flask-cors
pillow
rembg
onnxruntime
python-dotenv
numpy